+++ chartlets.js/CHANGES.md
@@ -37,8 +37,11 @@
   - using `schema` instead of `type` property for callback arguments
   - using `return` object with `schema` property for callback return values
 
-* New components
+* New (MUI) components
   - `LinearProgress`
+  - `Switch`
+
+* Supporting `tooltip` property for interactive MUI components.
 
 ## Version 0.0.29 (from 2024/11/26)